Basic RCON Receive functionality
The module can now receive everything except the body, which some might
say is the most important part but they can be quiet.

Multi-Packet Responses are still not implemented for now, but that shouldn't be
too much of a problem for the time being.

Random now has a proper seed, and shouldn't generate the same id for every command